requirements.txt
Django>=4.0.0,<5.0.0
djangorestframework>=3.13.0,<4.0.0
djangorestframework-simplejwt>=5.0.0,<6.0.0
django-cors-headers>=3.10.0,<4.0.0
djoser>=2.1.0,<3.0.0